backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adversarial-science"
version = "1.0.0"
description = "AI Tribunal for Scientific Paper Review"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "SpoonOS Hackathon Team"}
]
keywords = ["ai", "science", "tribunal", "neo", "blockchain"]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "spoon-ai-sdk>=0.3.3",
    "spoon-toolkits>=0.2.2",
    "anthropic>=0.18.0",
    "elevenlabs>=2.25.0",
    "pymupdf>=1.23.0",
    "neo-mamba>=0.11.0",
    "mem0ai>=0.0.20",
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]
contract = [
    "neo3-boa>=1.3.0",  # Requires Python 3.11-3.12
]

[project.scripts]
adversarial-science = "src.api.main:run"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"

[tool.ruff]
line-length = 100
target-version = "py310"
